Allow double values in recipe ingredient quantities

+++ Receptai-server/src/main/java/lt/tastybytes/receptaiserver/dto/recipe/IngredientDto.java
@@ -7,7 +7,7 @@ import org.springframework.lang.NonNull;
 public record IngredientDto(
         @NotEmpty @NonNull String name,
         @Positive
-        int quantity,
+        double quantity,
         @NotEmpty @NonNull String unit
 ) {
 }
